zsh: use zkbd for keybindings.
Instead of trying to set all of the keys manually which didn't work most
of the time, use zkbd.

To generate a new mapping:
autoload zkbd && zkbd

For any keys that cannot be pressed (like f11 or menu) use space to
skip.

The file will be stored in ~/.zkbd/${TERM}-${DISPLAY}, remove the
-${DISPLAY}